requirements.txt
amqp==2.6.0
anyjson==0.3.3
asgiref==3.2.7
billiard==3.5.0.5
celery==4.2.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.7
django-celery==3.3.1
django-celery-beat==2.0.0
django-crispy-forms==1.9.1
django-kombu==0.9.4
django-timezone-field==4.0
djangorestframework==3.11.0
future==0.18.2
httpie==2.1.0
idna==2.9
importlib-metadata==1.6.1
kombu==4.6.10
librabbitmq==2.0.0
msgpack==1.0.0
pkg-resources==0.0.0
pycparser==2.20
Pygments==2.6.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
